conditionally import types to avoid circular dependence
c.f. https://stackoverflow.com/a/39757388
